server/tests/requirements.txt
requests~=2.32.3
websocket-client~=1.8.0